# swagger.yaml
openapi: 3.0.0
info:
  title: Finance Dashboard API
  version: 1.0.0
  description: |
    Role-based finance dashboard backend built for Zorvyn's assessment.
    
    **Roles and access:**
    - `viewer` — dashboard endpoints only
    - `analyst` — read transactions + dashboard
    - `admin` — full access including user management
servers:
  - url: http://localhost:5000/api
    description: Local development
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
...
    Transaction:
      type: object
      properties:
        _id:
          type: string
        amount:
          type: number
        type:
          type: string
          enum: [income, expense]
        category:
          type: string
          enum: [salary, marketing, infrastructure, operations, revenue, investment, refund, tax, other]
        description:
          type: string
        date:
          type: string
          format: date
        createdBy:
          type: object
          properties:
            name:
              type: string
            email:
              type: string
...
security:
  - bearerAuth: []
paths:
...
  /users:
    get:
      tags: [Users]
      summary: List all users (admin only)
      responses:
        "200":
          description: List of users
        "403":
          description: Forbidden
...
  /transactions:
    get:
      tags: [Transactions]
      summary: List transactions with optional filters (analyst, admin)
      parameters:
        - in: query
          name: type
          schema:
            type: string
            enum: [income, expense]
        - in: query
          name: category
          schema:
            type: string
        - in: query
          name: startDate
          schema:
            type: string
            format: date
        - in: query
          name: endDate
          schema:
            type: string
            format: date
        - in: query
          name: page
          schema:
            type: integer
            default: 1
        - in: query
          name: limit
          schema:
            type: integer
            default: 10
      responses:
        "200":
          description: Paginated transaction list
        "403":
          description: Viewers cannot access raw transactions

    post:
      tags: [Transactions]
      summary: Create a new transaction (admin only)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/Transaction"
      responses:
        "201":
          description: Transaction created
        "422":
          description: Validation errors
...